pyproject.toml
[project]
name = "kcfinder-client"
version = "0.1.0"
description = "Python client library for KCFinder file manager"
readme = "README.md"
license = "MIT"
requires-python = ">=3.14"
dependencies = [
    "httpx>=0.28",
]

[project.urls]
Repository = "https://github.com/cyberops7/kcfinder-client"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/kcfinder_client"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.ruff]
src = ["src", "tests"]
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[dependency-groups]
dev = [
    "bandit",
    "invoke",
    "pip-audit",
    "pyrefly",
    "pytest",
    "pytest-asyncio",
    "pytest-httpx",
    "python-dotenv>=1.2.2",
    "ruff",
]
